+++ src/config.js
@@ -1,3 +1,3 @@
-const BASE_API = 'http://10.58.52.219:8000';
+const BASE_API = 'http://10.58.52.62:8000';
 
 export default BASE_API;
